Update monadic-parsers-angstrom.md
First of all, the parser monad is one monad out of many, so what is said is actually the specifics of the parser monad, which is a combination of (tri-)result and state monads. 

Also, the point that will produce lots of confusion - monad is not a function. Monad is a composition of computations. Basically, why mparsers use monads, because the allow you to declarativly combine you commands. It is a command combinator. Cf., IO Monad in Haskell.