Commented out links to additional missing or bad guides.
Commented out the Cornell guide and CIS500 guides.  It's a variant of the UPenn guide (or vice versa) and includes at least one of the same bits of SMLish non-OCaml code.  The CIS500 guide appears to be a copy of the UPenn guide, and also includes the bad code.

Also commented out the Pyrocat guide.  It's 404'ed, and Pyrocat doesn't seem to have the repo any longer.

I'm leaving the old URLs in comments for now in case anyone wants to check them without getting history of the page, but the comments should be deleted at some point. 

(But if the UPenn/Cornell guides seem good, maybe it would be worth copying one and fixing it.)