Removed commented-out broken pyrocat101 guide link
This is 404'ed:
Pyrocat OCaml Style Guide : https://github.com/pyrocat101/ocaml-style-guide
Pyrocat101 doesn't report his/her email in github, and it's not available via
a "find email addresses" via a ghithub api events url trick.  If anyone else knows what this guide
is, feel free to add it, but the line in awesome-ocaml that lists this guide seems to suggest that it
was just a variant of the Penn/Cornell guide.

By the way, about the three variants of the Penn/Cornell guide, I understand that people have found it/them helpful.  (All three look like they might have identical comment, but maybe there are differences.)  My personal feeling is that I don't trust a style guide with code that can't compile because it includes obviously incorrect syntax (see e.g. item 22).   It means that no one has gone through the guide carefully to make sure that what it says is appropriate. 

My hypothesis is that this was originally an SML style guide, and someone just did a quick search and replace to substitute "let" for "fun" at the beginnings of lines.  After that, professors copied it without reading through it carefully.  I wouldn't assume that an SML style guide is completely relevant to OCaml, though.  So if were up to me, I would not link to the Penn/Cornell guides at all.  (If someone were to make a version that fixes the problem, and checks that the guidelines really are appropriate to OCaml as it's used now, then let's use that.)  However, @pmetzger indicated on Discord that he's found the Penn/Cornell guide helpful, so I thought it best to leave the links but add the information that they are similar, along with a warning about the non-OCaml.