clarifies the staticness of the type checker
The main purpose of the "We can't say that this program is well-formed until someone puts a value of a wrong type into the cell." sentence was to highlight the static notion of the type. Due to the presence of dynamically typed languages, such as python and javascript, the idea that a type of a term should be preserved is not well understood and not absolutely natural. So, I believe, that we should give some hints to a reader, so that the right intuition could be built.